refactor reboot code to reboot-if-necessary.sh and autoupdate-and-reboot.sh to handle it. Finally, added code to silence meaningless errors around snapd

+++ remove-old-snaps.sh
@@ -3,11 +3,24 @@ PATH=/bin:/usr/sbin:/sbin:/usr/local/sbin
 # Removes old revisions of snaps
 # CLOSE ALL SNAPS BEFORE RUNNING THIS
 set -eu
-echo "refreshing snaps"
-snap refresh 
-echo "removing old snaps"
-LANG=en_US.UTF-8 snap list --all | awk '/disabled/{print $1, $3}' |
-    while read snapname revision; do
-        snap remove "$snapname" --revision="$revision"
-    done
-echo "done with snap cleanup"
+
+# Check if snap is installed
+if ! command -v snap &> /dev/null; then
+    echo "Snap is not installed. Skipping snap cleanup."
+    exit 0
+fi
+
+echo "Refreshing snaps"
+snap refresh
+
+# Check if there are any snaps installed
+if snap list --all | grep -q disabled; then
+    echo "Removing old snaps"
+    LANG=en_US.UTF-8 snap list --all | awk '/disabled/{print $1, $3}' |
+        while read snapname revision; do
+            snap remove "$snapname" --revision="$revision"
+        done
+    echo "Done with snap cleanup"
+else
+    echo "No old snaps to remove."
+fi
